Add test for addItem function
