+++ src/_mocks_/localStorage.js
@@ -1,9 +1,10 @@
 const store = [];
-const localstorageMock = {
+
+const localStorageMock = {
   getItem: (key) => store[key],
   setItem: (key, value) => {
     store[key] = value;
   },
 };
 
-export default localstorageMock;
+export default localStorageMock;
